pyproject.toml
[project]
name = "cocai"
version = "0.1.0"
description = "A chatbot that plays Call of Cthulhu (CoC) with you, powered by AI."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiosqlite>=0.20.0",
    "arize-phoenix>=4.32.0",
    "chainlit>=1.3.0rc0",
    "cochar>=1.0.1",
    "llama-index-agent-openai>=0.3.0",
    "llama-index-embeddings-ollama>=0.3.0",
    "llama-index-llms-ollama",
    "llama-index-llms-openai-like>=0.2.0",
    "llama-index-llms-openai>=0.2.0",
    "llama-index-tools-tavily-research>=0.2.0",
    "llama-index>=0.5.1",
    "rich>=13.8.0",
    "ruff>=0.6.3",
    "tomlkit>=0.13.2",
]
